feat(spell): implement complete spell module with 234-column mapping and field validation
- Add full 234 physical column mapping for foxy.dbc_spell entity and convert repeated slots to scalar fields
- Add validate() methods to spell area, custom attr, group, linked spell, loot template, and rank entities
- Align default values with core base SQL (gender=2, questStartStatus=64, questEndStatus=11, chance=100, lootMode=1, minCount=1, maxCount=1)
- Replace number inputs with entity pickers, flag pickers, and select dropdowns for better UX
- Remove unreliable copy functions that used MAX+1 (now throw UnsupportedError to prevent invalid references)
- Add missing fields: reagent slots 4-7, shapeshift mask/ exclude high 32 bits, exclude aura states/spells, language flags
- Refactor effect section to support dynamic titles and improve read-only logic
- Add comprehensive contract tests covering entity columns, default values, enums, validation, and copy errors

+++ lib/entity/spell_bonus_data_entity.dart
@@ -19,10 +19,10 @@ class SpellBonusDataEntity {
   factory SpellBonusDataEntity.fromJson(Map<String, dynamic> json) {
     return SpellBonusDataEntity(
       entry: json['entry'] ?? 0,
-      directBonus: (json['direct_bonus'] ?? 0.0),
-      dotBonus: (json['dot_bonus'] ?? 0.0),
-      apBonus: (json['ap_bonus'] ?? 0.0),
-      apDotBonus: (json['ap_dot_bonus'] ?? 0.0),
+      directBonus: (json['direct_bonus'] ?? 0).toDouble(),
+      dotBonus: (json['dot_bonus'] ?? 0).toDouble(),
+      apBonus: (json['ap_bonus'] ?? 0).toDouble(),
+      apDotBonus: (json['ap_dot_bonus'] ?? 0).toDouble(),
       comments: json['comments'] ?? '',
     );
   }
@@ -78,10 +78,10 @@ class BriefSpellBonusDataEntity {
   factory BriefSpellBonusDataEntity.fromJson(Map<String, dynamic> json) {
     return BriefSpellBonusDataEntity(
       entry: json['entry'] ?? 0,
-      directBonus: (json['direct_bonus'] ?? 0.0),
-      dotBonus: (json['dot_bonus'] ?? 0.0),
-      apBonus: (json['ap_bonus'] ?? 0.0),
-      apDotBonus: (json['ap_dot_bonus'] ?? 0.0),
+      directBonus: (json['direct_bonus'] ?? 0).toDouble(),
+      dotBonus: (json['dot_bonus'] ?? 0).toDouble(),
+      apBonus: (json['ap_bonus'] ?? 0).toDouble(),
+      apDotBonus: (json['ap_dot_bonus'] ?? 0).toDouble(),
       comments: json['comments'] ?? '',
     );
   }
